Export GenerateSignedLinkToken, do not set permission in CopyContentObject
